chore: extract from_cli_for_struct module
