+++ interactive-clap-derive/src/derives/interactive_clap/common_methods/from_cli_for_enum.rs
