+++ interactive-clap-derive/src/derives/interactive_clap/common_methods/mod.rs
@@ -3,7 +3,6 @@ pub mod fields_with_skip_default_input_arg;
 pub mod fields_with_subargs;
 pub mod fields_with_subcommand;
 pub mod from_cli_for_enum;
-pub mod from_cli_for_struct;
 pub mod input_arg;
 pub mod interactive_clap_attrs_context;
 pub mod skip_interactive_input;
