+++ interactive-clap-derive/src/derives/interactive_clap/structs/from_cli_trait.rs
@@ -5,14 +5,14 @@ use proc_macro_error::abort_call_site;
 use quote::{quote, ToTokens};
 use syn;
 
-pub fn from_cli_for_struct(
-    ast: &syn::DeriveInput,
-    fields: &syn::Fields,
-) -> proc_macro2::TokenStream {
+use crate::derives::interactive_clap::common_methods;
+
+/// returns the whole result `TokenStream` of derive logic of containing module
+pub fn token_stream(ast: &syn::DeriveInput, fields: &syn::Fields) -> proc_macro2::TokenStream {
     let name = &ast.ident;
 
     let interactive_clap_attrs_context =
-        super::interactive_clap_attrs_context::InteractiveClapAttrsContext::new(ast);
+        common_methods::interactive_clap_attrs_context::InteractiveClapAttrsContext::new(ast);
     if interactive_clap_attrs_context.is_skip_default_from_cli {
         return quote!();
     };
@@ -20,8 +20,8 @@ pub fn from_cli_for_struct(
     let fields_without_subcommand_and_subargs = fields
         .iter()
         .filter(|field| {
-            !super::fields_with_subcommand::is_field_with_subcommand(field)
-                && !super::fields_with_subargs::is_field_with_subargs(field)
+            !common_methods::fields_with_subcommand::is_field_with_subcommand(field)
+                && !common_methods::fields_with_subargs::is_field_with_subargs(field)
         })
         .map(|field| {
             let ident_field = &field.clone().ident.expect("this field does not exist");
@@ -102,7 +102,7 @@ fn fields_value(field: &syn::Field) -> proc_macro2::TokenStream {
     let ident_field = &field.clone().ident.expect("this field does not exist");
     let fn_input_arg = syn::Ident::new(&format!("input_{}", &ident_field), Span::call_site());
     if field.ty.to_token_stream().to_string() == "bool"
-        || super::skip_interactive_input::is_skip_interactive_input(field)
+        || common_methods::skip_interactive_input::is_skip_interactive_input(field)
     {
         quote! {
             let #ident_field = clap_variant.#ident_field.clone();
@@ -123,8 +123,8 @@ fn fields_value(field: &syn::Field) -> proc_macro2::TokenStream {
             };
             let #ident_field = clap_variant.#ident_field.clone();
         }
-    } else if !super::fields_with_subcommand::is_field_with_subcommand(field)
-        && !super::fields_with_subargs::is_field_with_subargs(field)
+    } else if !common_methods::fields_with_subcommand::is_field_with_subcommand(field)
+        && !common_methods::fields_with_subargs::is_field_with_subargs(field)
     {
         quote! {
             if clap_variant.#ident_field.is_none() {
